refactor: decouple searchbar from wibarab data structure
